Imports [1/n]
- Basic definitions for handling imports
- Definitions of exported declarations were extracted to a separate 
file, so they can be reused for imports
