+++ src/Frontend/Ast.hs
@@ -1,13 +1,13 @@
 {- |
-Module      :  Frontend.Export.Ast
-Description :  Result of processing of a module
+Module      :  Frontend.Ast
+Description :  Declarations for imports and exports
 Copyright   :  (c) Danil Kolikov, 2019
 License     :  MIT
 
-AST of the result of the processing of a module
+Declarations, used for imports and exports
 -}
-module Frontend.Export.Ast
-    ( module Frontend.Export.Ast
+module Frontend.Ast
+    ( module Frontend.Ast
     , module Core.Ident
     , module Frontend.Desugaring.Final.Ast
     , module Frontend.Inference.Instance
@@ -20,28 +20,6 @@ import Frontend.Desugaring.Final.Ast (Fixity(..))
 import Frontend.Inference.Constraint
 import Frontend.Inference.Instance
 import Frontend.Inference.Signature
-import Util.HashMap
-
--- | A result of module processing
-data Module = Module
-    { getModuleDataTypes :: DataTypes
-    , getModuleTypeSynonyms :: Signatures TypeSignature
-    , getModuleClasses :: Classes
-    , getModuleInstances :: Instances
-    , getModuleExpressions :: Expressions
-    } deriving (Eq, Show)
-
-instance Semigroup Module where
-    Module d1 t1 c1 i1 e1 <> Module d2 t2 c2 i2 e2 =
-        Module
-            (deepMerge d1 d2)
-            (t1 <> t2)
-            (deepMerge c1 c2)
-            (i1 <> i2)
-            (e1 <> e2)
-
-instance Monoid Module where
-    mempty = Module mempty mempty mempty mempty mempty
 
 -- | Definition of a data type
 data DataType = DataType
@@ -65,9 +43,6 @@ data Constructor = Constructor
 -- | A mapping from the name of a field to its position in a constructor
 type FieldsMap = HashMap Ident Int
 
--- | A map of data types
-type DataTypes = HashMap Ident DataType
-
 -- | Definition of a type class
 data Class = Class
     { getClassContext :: [SimpleConstraint] -- ^ Context of a type class
@@ -85,20 +60,11 @@ instance Semigroup Class where
             then Class c1 s2 d1 co1 (g1 <> g2) (m1 <> m2) di2
             else error "Can't merge classes with different signatures"
 
--- | A map of classes
-type Classes = HashMap Ident Class
-
--- | A map of defined instances
-type Instances = HashMap Ident Instance
-
 -- | Definition of an expression
 data Expression = Expression
     { getExpressionType :: TypeSignature -- ^ Optional type signature
     , getExpressionFixity :: Maybe FixitySignature -- ^ Fixity of an expression
     } deriving (Eq, Show)
-
--- | A map of expressions
-type Expressions = HashMap Ident Expression
 
 -- | A fixity of an expression
 data FixitySignature = FixitySignature
